Implement loading.tsx for dynamic paths.
Learned about static and dynamic rendering.
loading.tsx is a nextjs comp built on top of React's <Suspense>
loading.tsx will be shown while server components (admin/products in this case)
are busy fetching data.
Server comps can only be shipped to the client when all the data fetching is done,
loading.tsx is used to display something in the meanwhile, letting the user know webpage is "loading".